fix: P2 批次二——时区重载顺序、分页溢出、vN 刷新 URL、用量缓存竞态与单飞
- 11-01：设置页改时区时 reload_all_jobs 在落库前执行，调度器按旧时区重算
  next_run_at（与注释承诺相反）。移到落库 + 缓存刷新之后，对齐备份导入路径。
- 11-03：request_logs offset 用 u32 相乘，page > ~4294 万时 dev panic /
  release 静默回绕返回错页。改先转 i64 再乘。
- 14-01：build_models_url 版本段只认 v1/v1beta/v1alpha，v3/v4 末段被当普通
  路径再补一层 /v1（9 个种子供应商刷新模型必 404）。判定收敛为单实现
  provider_model::refresh::is_version_segment，与转发侧 build_upstream_url 共用；
  14-05 反向注释一并修正并补 vN 测试。
- 11-02 + 11-25：用量缓存失效与在途抓取无护栏、读端点抓取无单飞。UsageMemCache
  增加 provider 失效代次（invalidate 自增），单飞抓取开始前记录、写库/回填前
  比对，期间失效即作废结果（UsageError::Stale，可重试）；/usage 与
  /usage/estimate 改走 fetch_shared_stored，与 LB 兜底共用同一单飞入口。
  单测含先红后绿的 mid-flight 失效回归。ADR-0009 Decision 5。

质量门：cargo fmt --check / clippy -D warnings / cargo test --all-targets
（851 passed）全绿。

+++ src/routes/request_logs.rs
@@ -133,7 +133,8 @@ async fn list_request_logs(
         .page_size
         .unwrap_or(DEFAULT_PAGE_SIZE)
         .clamp(1, MAX_PAGE_SIZE);
-    let offset = ((page - 1) * page_size) as i64;
+    // 先转 i64 再乘：u32 相乘在 page > ~4294 万时溢出（dev panic / release 静默回绕）。
+    let offset = (i64::from(page) - 1) * i64::from(page_size);
 
     // 拼接 WHERE 条件与绑定参数。
     let mut where_sql = String::from("WHERE 1=1");
